pyproject.toml
[project]
name = "tricked"
version = "0.1.0"
description = "Native Rust Tricked AI Engine"
dependencies = [
    "numpy>=1.24.0",
    "pyzmq>=24.0.0",
    "hydra-core>=1.3.0",
    "wandb>=0.15.0",
    "redis>=4.5.0",
    "rich>=13.0.0",
    "ray>=2.9.0",
    "lz4>=4.3.0",
    "torch>=2.4.0"
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
    "maturin>=1.12"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff.lint]
ignore = ["E402"]
